==> map_matching/s1_map_Nodes.ipynb <==
from bs4 import BeautifulSoup as bs
import requests as req
from pprint import *
bsr = bs(open("../data/jiading_map/Jiading.osm"),"xml")

nodes = bsr.find_all('node')
print(len(nodes))
# print nodes

import pandas as pd
datalist=[]
for node in nodes:
    datalist.append([node['id'],node['lat'],node['lon']])
df1=pd.DataFrame(datalist,columns=['origin_id','lat','lon'])
df1['new_id']=df1.index
df1

print df1.lat.max()
print df1.lat.min()

print df1.lon.max()
print df1.lon.min()

df1.to_csv('../data/jiading_map/Nodes.csv',index=False)

# None sense...
# length=len(df1)
# file=open('WA_Nodes.txt','wb')
# for i in range(length):
#     print>>file,    df1.loc[i,'new_id'],df1.loc[i,'lat'],df1.loc[i,'lon']

# file.close()

ways = bsr.find_all('way')
way_kinds = ['motorway','motorway_link','primary','primary_link',\
            'secondary','secondary_link','tertiary', 'tertiary_link','residential','living_street',\
            'service','trunk','trunk_link','unclassified','road','track']

highways = []
way_types = set(['pedestrian', 'footway', 'steps'])
for w in ways:
    flag=False
    way_id = int(w.attrs['id'])
    for tag in w.find_all('tag'):
        if tag['k']=='highway':
            if tag['v'] in way_kinds and way_id not in [266205933, 266205935]:
                flag=True
                break
            else:
                if tag['v'] in way_types and way_id in [384531666, 384531667, 266221576, 263913992, 263913994, 135419990, 135372842, 135372843, 135372844, 135372845]:
                    flag=True
                    break
                
    if flag:
        highways.append(w)
# highways[:5]
print way_types, len(highways)

import pandas as pd
from collections import defaultdict
segMap = defaultdict(int)

def chunks(l, n):
    """Yield successive n-sized chunks from l."""
    for i in range(0, len(l), n-1):
        yield l[i:i+n]
final_list=[]
for w in highways:

    isOneway = False
    category = 'unclassified'
    for tag in w.find_all('tag'):
        if tag['k']=='oneway' and tag['v'] =='yes':
            isOneway = True
        if tag['k']=='highway':
            category = tag['v']

    nodes=[nd['ref'] for nd in w.find_all('nd')]
    for node in nodes:
        segMap[node] += 1
    len_nodes=len(nodes)
    final_list.append([1, False,  category,nodes])
#     final_list.append([1, False,  category,nodes[::-1]])
#         nodes_list=list(chunks(nodes,4))
#         if(len(nodes_list[-1])==1):
#             nodes_list.pop()    
#         for nds in nodes_list:
#             final_list.append([1, isOneway,  category,nds])
junctions = set()       
for k in segMap.keys():
    if segMap[k] > 1:
        junctions.add(k)
    
print len(junctions)
df = pd.DataFrame(final_list,columns=['id','oneway','category','nodes'])
df

a = [1759597604, 3878436154, 2703796554]
list(reversed(a))

final_list=[]
for index, row in df.iterrows():
#     print index, row['nodes']
    st = 0;
    node_list = []
    for i in range(1, len(row['nodes']) - 1):
        if row['nodes'][i] in junctions:
            node_list.append(row['nodes'][st:i+1])
            node_list.append(list(reversed(row['nodes'][st:i+1])))
            st = i
    node_list.append(row['nodes'][st:len(row['nodes'])])
    node_list.append(list(reversed(row['nodes'][st:len(row['nodes'])])))
    for nst in node_list:
        final_list.append([1, row['oneway'], row['category'], nst]) 
df = pd.DataFrame(final_list,columns=['id','oneway','category','nodes'])
df

df['id']=df.index
df['nodes']=df['nodes'].apply(lambda x:list(x))
df

df.to_csv('../data/jiading_map/highway.csv',index=False)
