==> map_matching/s2_map_Edges.ipynb <==
import pandas as pd
df=pd.read_csv('../data/jiading_map/Nodes.csv')
df

def get_node_lat_lon(node_id):

    node=df[df.origin_id==node_id]
    return node.iloc[0,1],node.iloc[0,2]

# print get_node_lat_lon(4727003825)


def cal_distance(nodeA,nodeB):
#     nodeA=eval(nodeA)
#     nodeB=eval(nodeB)
    from geopy.distance import vincenty
    latA,lonA=get_node_lat_lon(nodeA)
    latB,lonB=get_node_lat_lon(nodeB)
    a=(latA,lonA)
    b=(latB,lonB)

    distance=vincenty(a, b).m
    return distance
    
# print cal_distance(4727003825,4727003824)

def distance(strlist):
    list=[eval(node_id) for node_id in eval(strlist)]
    
    dis=0
    for i in range(len(list)-1):
        dis=dis+cal_distance(list[i],list[i+1])
    return dis

ddf=pd.read_csv('../data/jiading_map/highway.csv')[['id','nodes']]
# print ddf
ddf['from']=ddf['nodes'].apply(lambda x:eval(x)[0])
ddf['to']=ddf['nodes'].apply(lambda x:eval(x)[-1])

# from __future__ import division  
# ddf['distance']=ddf['nodes'].apply(lambda x: distance(x))
# ddf['speed']=(ddf['maxspeed']+ddf['minspeed'])/2
ddf['cost']=ddf['nodes'].apply(lambda x:distance(x))
ddf

ddff=ddf[['id','from','to','cost','nodes']] 
ddff.to_csv('../data/jiading_map/Edges.csv',index=False)
ddff.head()

import pandas as pd
df=pd.read_csv('../data/jiading_map/Nodes.csv')
ddff=pd.read_csv('../data/jiading_map/Edges.csv')
def get_newid(origin_id):
    return int(df[df.origin_id==origin_id].loc[:,'new_id'])
    
# print get_newid(25632292)
ddff['from']=ddff['from'].apply(lambda x:get_newid((x)))
ddff['to']=ddff['to'].apply(lambda x:get_newid((x)))
ddff.head()

ddff[['id','from','to','cost']].to_csv('../data/jiading_map/WA_Edges.csv',index=False)
import pandas as pd
# newdf=pd.read_csv('./data/jiading_map/WA_Edges.csv')
# newdf['cost']=newdf['cost'].apply(lambda x:1)
# length=len(newdf)
# file=open('WA_Edges.txt','wb')
# for i in range(length):
#     print>>file,    int(newdf.loc[i,'id']),int(newdf.loc[i,'from']),int(newdf.loc[i,'to']),int(newdf.loc[i,'cost'])

# file.close()
