==> map_matching/s3_map_EdgeGeometry.ipynb <==
import pandas as pd
highway_df=pd.read_csv('../data/jiading_map/highway.csv')
edge_df=pd.read_csv('../data/jiading_map/Edges.csv')
print len(highway_df), len(edge_df)
result=pd.merge(highway_df,edge_df,on=['id','nodes'])
# alldata_df['length']=alldata_df['distance']*1.609344*1000
result

new=result[['id','category','nodes']]
# new['distance']=new['distance']*1.609344*1000
# new['distance']=new['category'].apply(lambda x:1)
new['distance'] = result['cost']
new

new['category'].unique()

import pandas as pd
f=pd.read_csv('../data/jiading_map/Nodes.csv')

def get_node_lat_lon(node_id):
#     if (node_id.contains("'")):
#         node_id=node_id[1:-1]
    node=f[f.origin_id==node_id]
    return str(node.iloc[0,1])+'^'+str(node.iloc[0,2])

# get_node_lat_lon(4704039842)

def format(strlist):
    list=[eval(node_id) for node_id in eval(strlist)]
    
    str=''
    for i in range(len(list)):
        str=str+'^'+get_node_lat_lon(list[i])
    return str

new['strnodes']=new['nodes'].apply(lambda x: format(x))
new

new[['id','category','distance','nodes','strnodes']].to_csv('../data/jiading_map/WA_EdgeGeometry.csv',index=False)

import pandas as pd
new=pd.read_csv('../data/jiading_map/WA_EdgeGeometry.csv')
# new['distance']=new['distance'].apply(lambda x:1)
length=len(new)
file=open('../data/jiading_map/jiading_EdgeGeometry.txt','wb')
for i in range(length):
    print>>file,    str(int(new.loc[i,'id']))+'^^'+str(new.loc[i,'category'])+'^'+str(int(new.loc[i,'distance']))+str(new.loc[i,'strnodes'])

file.close()
