==> map_matching/s4_map_Clean.ipynb <==
import pandas as pd
df = pd.read_csv('../data/jiading_map/Edges.csv')
df

length=len(df)
# press_nodes=set()
press_nodes = set(df['from']) | set(df['to'])
# for i in range(length):
    
#     nodes=eval(df.iloc[i,4])
#     nodes_list=[nodes[0],nodes[-1]]
#     for node in  nodes_list:
#         press_nodes.add(node)
    
# print press_nodes
len(press_nodes)

press_nodes=[int(node) for node in press_nodes]
ndf=pd.read_csv('../data/jiading_map/Nodes.csv')
new_ndf=ndf[ndf['origin_id'].isin(press_nodes)]
new_ndf

print new_ndf.index
print len(new_ndf.index)

index=pd.Series(range(len(new_ndf.index)))
new_ndf.index=index
new_ndf['new_id']=index
new_ndf


length=len(new_ndf)
file=open('../data/jiading_map/jiading_Nodes.txt','wb')
for i in range(length):
    print>>file,    int(new_ndf.loc[i,'new_id']),new_ndf.loc[i,'lat'],new_ndf.loc[i,'lon']

file.close()

def get_newid(origin_id):
    return int(new_ndf[new_ndf.origin_id==origin_id].loc[:,'new_id'])
df['from']=df['from'].apply(lambda x:get_newid((x)))
df['to']=df['to'].apply(lambda x:get_newid((x)))
df

length=len(df)
file=open('../data/jiading_map/jiading_Edges.txt','wb')
for i in range(length):
    print>>file,    int(df.loc[i,'id']),int(df.loc[i,'from']),int(df.loc[i,'to']),float(df.loc[i,'cost'])

file.close()
